advert

+++ src/main/java/com/project/payload/response/business/CityResponse.java
@@ -1,6 +1,5 @@
 package com.project.payload.response.business;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -14,5 +13,5 @@ public class CityResponse {
 
     private Long id;
     private String name;
-    private int countryId;
+    private Long countryId;
 }
